# upgrade to webpack v4