--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -18,7 +18,7 @@ jobs:
 
       - run:
           name: Install Yarn
-          command: sudo npm install -g yarn@~1.0.0
+          command: sudo npm install -g yarn@^1.0.0
 
       - run:
           name: Install Project Dependencies