--- a/package.json
+++ b/package.json
@@ -61,7 +61,7 @@
     "url-loader": "^1.0.0-beta.0",
     "webpack": "^4.0.0",
     "webpack-cli": "^2.0.8",
-    "webpack-dev-server": "^3.0.0-alpha6",
+    "webpack-dev-server": "^3.0.0",
     "webpack-md5-hash": "^0.0.5",
     "webpack-merge": "^4.1.2",
     "webpack-pwa-manifest": "^3.3.2",