--- a/package.json
+++ b/package.json
@@ -44,7 +44,7 @@
     "extract-text-webpack-plugin": "^4.0.0-alpha.0",
     "favicons-webpack-plugin": "^0.0.7",
     "file-loader": "^1.1.9",
-    "html-critical-webpack-plugin": "^1.0.1",
+    "html-critical-webpack-plugin": "^1.1.0",
     "html-webpack-plugin": "webpack-contrib/html-webpack-plugin#a8a8c2",
     "http-server": "^0.10.0",
     "jest": "^21.2.1",
@@ -63,7 +63,7 @@
     "webpack-dev-server": "^3.0.0",
     "webpack-md5-hash": "^0.0.5",
     "webpack-merge": "^4.1.2",
-    "webpack-pwa-manifest": "^3.3.2",
+    "webpack-pwa-manifest": "^3.5.0",
     "webpack-s3-plugin": "^1.0.0-rc.0"
   }
 }