--- a/package.json
+++ b/package.json
@@ -56,12 +56,10 @@
     "sass-loader": "^6.0.6",
     "style-loader": "^0.20.2",
     "tapable": "^1.0.0",
-    "uglify-js": "git://github.com/mishoo/UglifyJS2#harmony-v2.8.22",
     "url-loader": "^1.0.0-beta.0",
     "webpack": "^4.0.0",
     "webpack-cli": "^2.0.8",
     "webpack-dev-server": "^3.0.0",
-    "webpack-md5-hash": "^0.0.5",
     "webpack-merge": "^4.1.2",
     "webpack-pwa-manifest": "^3.5.0",
     "webpack-s3-plugin": "^1.0.0-rc.0"