--- a/src/vendor.js
+++ b/src/vendor.js
@@ -1,7 +1,0 @@
-import 'jquery';
-import 'moment';
-import 'react';
-import 'react-dom';
-import 'react-lazyload';
-import 'react-router';
-import 'react-transition-group';