--- a/webpack.config.common.js
+++ b/webpack.config.common.js
@@ -42,7 +42,7 @@ module.exports = {
       }, {
         test: /\.woff(2)?(\?v=[0-9]\.[0-9]\.[0-9])?$/,
         loader: 'url-loader?limit=10000&mimetype=application/font-woff'
-      },{
+      }, {
         test: /\.(ttf|eot|svg|jpe?g|png|gif|otf)(\?v=[0-9]\.[0-9]\.[0-9])?$/,
         loader: 'file-loader'
       }