--- a/webpack.config.common.js
+++ b/webpack.config.common.js
@@ -22,32 +22,28 @@ module.exports = {
   },
 
   module: {
-    // TODO - https://github.com/webpack-contrib/eslint-loader/pull/212
-    rules: [
-      // {
-      //   test: /\.(js*)x$/,
-      //   enforce: 'pre',
-      //   loader: 'eslint-loader',
-      //   exclude: path.join(__dirname, 'node_modules')
-      // },
-      {
-        test: /\.(js*)x$/,
-        loaders: [
-          'babel-loader',
-          'react-hot-loader/webpack'
-        ],
-        exclude: path.join(__dirname, 'node_modules')
-      }, {
-        test: /\.(s*)css$/,
-        use: ['style-loader', 'css-loader', 'sass-loader']
-      }, {
-        test: /\.woff(2)?(\?v=[0-9]\.[0-9]\.[0-9])?$/,
-        loader: 'url-loader?limit=10000&mimetype=application/font-woff'
-      }, {
-        test: /\.(ttf|eot|svg|jpe?g|png|gif|otf)(\?v=[0-9]\.[0-9]\.[0-9])?$/,
-        loader: 'file-loader'
-      }
-    ]
+    rules: [{
+      test: /\.(js*)x$/,
+      enforce: 'pre',
+      loader: 'eslint-loader',
+      exclude: path.join(__dirname, 'node_modules')
+    }, {
+      test: /\.(js*)x$/,
+      loaders: [
+        'babel-loader',
+        'react-hot-loader/webpack'
+      ],
+      exclude: path.join(__dirname, 'node_modules')
+    }, {
+      test: /\.(s*)css$/,
+      use: ['style-loader', 'css-loader', 'sass-loader']
+    }, {
+      test: /\.woff(2)?(\?v=[0-9]\.[0-9]\.[0-9])?$/,
+      loader: 'url-loader?limit=10000&mimetype=application/font-woff'
+    }, {
+      test: /\.(ttf|eot|svg|jpe?g|png|gif|otf)(\?v=[0-9]\.[0-9]\.[0-9])?$/,
+      loader: 'file-loader'
+    }]
   },
 
   plugins: [