--- a/webpack.config.prod.js
+++ b/webpack.config.prod.js
@@ -1,6 +1,6 @@
 const commonConfig = require('./webpack.config.common');
-// const FaviconsWebpackPlugin = require('favicons-webpack-plugin');
-// const HtmlCriticalPlugin = require('html-critical-webpack-plugin');
+const FaviconsWebpackPlugin = require('favicons-webpack-plugin');
+const HtmlCriticalPlugin = require('html-critical-webpack-plugin');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin');
 const path = require('path');
@@ -38,29 +38,28 @@ module.exports = webpackMerge(commonConfig, {
   plugins: [
 
     // TODO - https://github.com/jantimon/favicons-webpack-plugin/issues/92
-    // new FaviconsWebpackPlugin({
-    //   logo: './components/bootstrap/images/pvd-geeks-logo.png',
-    //   emitStats: true,
-    //   prefix: 'icons/',
-    //   statsFilename: 'icons/stats.json',
-    //   inject: true,
-    //   title: 'Providence Geeks',
-    //   background: '#bcbfc2',
-    //   icons: {
-    //     android: true,
-    //     appleIcon: true,
-    //     appleStartup: true,
-    //     coast: false,
-    //     favicons: true,
-    //     firefox: true,
-    //     opengraph: true,
-    //     twitter: true,
-    //     yandex: true,
-    //     windows: true
-    //   }
-    // }),
+    new FaviconsWebpackPlugin({
+      logo: './components/bootstrap/images/pvd-geeks-logo.png',
+      emitStats: true,
+      prefix: 'icons/',
+      statsFilename: 'icons/stats.json',
+      inject: true,
+      title: 'Providence Geeks',
+      background: '#bcbfc2',
+      icons: {
+        android: true,
+        appleIcon: true,
+        appleStartup: true,
+        coast: false,
+        favicons: true,
+        firefox: true,
+        opengraph: true,
+        twitter: true,
+        yandex: true,
+        windows: true
+      }
+    }),
 
-    // TODO - https://github.com/arthurbergmz/webpack-pwa-manifest/issues/52
     new WebpackPwaManifest({
       name: 'Providence Geeks',
       short_name: 'PVD Geeks', // eslint-disable-line camelcase
@@ -76,20 +75,19 @@ module.exports = webpackMerge(commonConfig, {
       }]
     }),
 
-    // TODO - https://github.com/ProvidenceGeeks/website-frontend/pull/142
-    // new HtmlCriticalPlugin({
-    //   base: path.resolve(__dirname, 'build'),
-    //   src: 'index.html',
-    //   dest: 'index.html',
-    //   inline: true,
-    //   minify: true,
-    //   extract: true,
-    //   width: 375,
-    //   height: 565,
-    //   penthouse: {
-    //     blockJSRequests: false
-    //   }
-    // }),
+    new HtmlCriticalPlugin({
+      base: path.resolve(__dirname, 'build'),
+      src: 'index.html',
+      dest: 'index.html',
+      inline: true,
+      minify: true,
+      extract: true,
+      width: 375,
+      height: 565,
+      penthouse: {
+        blockJSRequests: false
+      }
+    }),
 
     new MiniCssExtractPlugin(),