<!--
## Submitting a Pull Request
We love PRs and appreciate any help you can offer!

Please make sure the following criteria are met before submitting your pull request.

1. <strong>PR meets the Contributing Guidelines (see above)</strong>
2. Ensure the test suite passes
3. Make sure your code lints
-->

## Related Issue
<!-- Include a link to the issue (e.g. #12) -->
resolves #141  

bonus: was able to resolve #139 as well.

## Summary of Changes
<!-- Briefly summarize the changes made, lists are also appreciated.  Referencing the related issue as a
"TO DO" checklist is also helpful for reviewers

1. [x] Thing I fixed
1. [x] Other thing I updated
1. [x] Docs I updated

-->
Upgrade to webpack v4
1.  Updated **webpack**, added **webpack-cli**, using contrib version of **html-webpack-plugin**, upgraded dependencies as needed
    * Removed `CommonsChunkPlugin`
    * Removed _vendor.js_ and implemented `optimization` configuration
    * Enabled `development` and `production` modes
    * Enabled `optimization` configuration
    * Was able to remove `DefinePlugin`
    * swapped ETWP with mini-css-extract-plugin
1. Upgraded build to NodeJS v.8.9.4 (LTS) per recommendation from webpack team
1. Cleaned up some of the `devDependencies`

### Stats
- Before: 99.19s - https://circleci.com/gh/ProvidenceGeeks/website-frontend/299#build-timing/containers/0
- Now: 35.07s - https://circleci.com/gh/ProvidenceGeeks/website-frontend/308#build-timing/containers/0

So could be potentially be a 50% savings even with another 15s of headroom.  pretty nice!  🎆 🎉 

But obviously without `ExtractTextWebpackPlugin`, our JS bundles are rather large
```shell
WARNING in asset size limit: The following asset(s) exceed the recommended size limit (244 KiB).
This can impact web performance.
Assets:
  vendor.49f48c8be89481dee508.bundle.js (478 KiB)
  index.5041108cb33b5dacaa6e.bundle.js (1.87 MiB)

WARNING in entrypoint size limit: The following entrypoint(s) combined asset size exceeds the recommended limit (244 KiB). This can impact web performance.
Entrypoints:
  index (1.87 MiB)
      index.5041108cb33b5dacaa6e.bundle.js
  vendor (478 KiB)
      vendor.49f48c8be89481dee508.bundle.js


WARNING in webpack performance recommendations:
You can limit the size of your bundles by using import() or require.ensure to lazy load some parts of your application.
For more info visit https://webpack.js.org/guides/code-splitting/
```

## TODO
1. 1. [x] [Had to install `tapable` locally to appease fork of html-weboack-plugin](https://github.com/webpack-contrib/html-webpack-plugin/issues/13).  **Update**: resolved in [webpack-contrib/html-webpack-plugin](https://github.com/webpack-contrib/html-webpack-plugin/issues/16#issuecomment-368616834) . **Update**: upstream has released a compat version
1. 1. [x] Had to comment out `eslint-loader`, [it's not supported with v4](https://github.com/webpack-contrib/eslint-loader/pull/212).  **Update**: resolved in [2.0.0 release](https://github.com/webpack-contrib/eslint-loader/pull/212#issuecomment-368447770)
1. 1. [x] Had to disable `ExtractTextWebpackPlugin`, [it is awaiting support for v4](https://github.com/webpack-contrib/extract-text-webpack-plugin/issues/701).  Trying out `MiniCssExtractPlugin` - ran into this issue https://github.com/webpack-contrib/mini-css-extract-plugin/issues/50
1. 1. [x] Had to disable `FaviconsWebpackPlugin`, [it needs help getting support for v4](https://github.com/jantimon/favicons-webpack-plugin/issues/92)
1. 1. [x] Had to disable `WebpackPwaManifest` plugin, [I opened an issue inquiring about support for v4](https://github.com/arthurbergmz/webpack-pwa-manifest/issues/52) .  **Update**: this [issue](https://github.com/arthurbergmz/webpack-pwa-manifest/issues/44#issuecomment-366456547) seems to be tracking the release.  
1. 1. [x] Had to disable `HtmlCriticalWebpackPlugin`, likely because of `ExtractTextWebpackPlugin` [but opened a PR to inquire about general webpack v4 support needed](https://github.com/anthonygore/html-critical-webpack-plugin/issues/9).  [Opened a PR](https://github.com/anthonygore/html-critical-webpack-plugin/pull/12) to upgrade the plugin to webpack v4
1. 1. [x] `OptimizeCssAssetsPlugin` left in but not sure if it's not running because `ExtractTextWebpackPlugin` is not enabled, [but opened a PR to inquire about general webpack v4 support needed](https://github.com/NMFR/optimize-css-assets-webpack-plugin/issues/34).