@thescientist13 2018-02-27
Hey @FDiskas , just a heads up I think you might have meant to comment in a different PR (for **webpack-contrib/html-webpack-plugin**)?  Or if you did mean this one, could you please elaborate?  Just keep in mind this PR for upgrading our application is still heavily WIP (work in progress) though I have been documenting it extensively so as to provide a guide to others doing the same. 

Hope that helps you out ✌️