@thescientist13 2018-03-23
## Overview
Final stats, all running locally (so no gzip, HTTP/2 etc) for getting a general feel of what things look like before / after the upgrade.

## Before
#### Build Time: 29.08s
```
                                  Asset      Size  Chunks             Chunk Names
    bbf1e3879689c911d25a8ebddb4d8f76.otf  13.9 KiB          [emitted]
    bc517149950afc2999fdec6f2e1bf064.otf  14.3 KiB          [emitted]
    Entrypoint mini-css-extract-plugin = *
       5 modules
✨  Done in 29.08s.
```
#### Network Requests: 17 (2 JS, 2 CSS)
#### Downloaded: 1.2MB (JS == 622 KB, CSS == 71.4 KB) 
#### Finished: 534ms
#### Lighthouse score: 52 / 55 / 76 / 88 / 89
<img width="760" alt="screen shot 2018-03-23 at 6 27 51 pm" src="https://user-images.githubusercontent.com/895923/37855872-f28c1b26-2ec7-11e8-9621-4e31b3ff931c.png">

## After
#### Build Time: 44.71s
```
     2 assets
       4 modules
✨  Done in 44.71s.
```
#### Network Requests: 17 (3 JS, 2 CSS)
#### Downloaded:  1.3 MB (JS = 716 KB, CSS = 71.4 KB)
#### Finished: 918ms
#### Lighthouse score: 47 / 55 / 76 / 88 / 89
<img width="710" alt="screen shot 2018-03-23 at 6 38 02 pm" src="https://user-images.githubusercontent.com/895923/37856171-67b06a5a-2ec9-11e8-99fe-4097bdd2736a.png">

## Summary
All metrics improved through this upgrade in particular build time 
just running locally, deploying this in the wild should bring even more improvements !!  🎉 

#### Build Time: down 15.63s (34%!)
#### Network Requests: -
#### Downloaded:  0.1MB (13%)
#### Finished: 384ms (41%)
#### Lighthouse score: 5 in performance (9.4%)

webpack is _really_ packing that JavaScript 🏆 

cc: @DevLab2425