@thescientist13 2018-03-27
Stage Before
<img width="740" alt="lighthouse-before" src="https://user-images.githubusercontent.com/895923/37985519-f582d458-31c6-11e8-8e96-aa0f4492e725.png">

<img width="1463" alt="network-tab-before-all" src="https://user-images.githubusercontent.com/895923/37985534-faf31236-31c6-11e8-98b0-1de8f939c095.png">
<img width="1526" alt="network-tab-before-css" src="https://user-images.githubusercontent.com/895923/37985535-fafe6f32-31c6-11e8-9e05-4a4e7d5ff17d.png">
<img width="1524" alt="network-tab-before-js" src="https://user-images.githubusercontent.com/895923/37985537-fb0b0ca6-31c6-11e8-9ad6-a1eda054f9b2.png">