@DevLab2425 2018-03-30
@thescientist13 
Just finding the time to dig in, and first let me say I'm no Webpack expert, nor a PWA expert, I just play one on ~TV~ a daily basis. 

Looking at your prod build I see a few slight differences from mine, but the majority looks similar to mine. (Considering I used this repo as a Webpack reference, there's no surprise there.)

A few things I have, that you do not:
- I'm using multiple entry files, you have a single. This gives me smaller assets in hopes of avoiding blocking scripts
- I have a minified CSS file injected into index.html. You seem to still have the [multiple CSS file issue](https://github.com/ProvidenceGeeks/website-frontend/issues/139)

A few thing you have, that I do not:
- The latest version (0.0.8) of [favicons-webpack-plugin](https://github.com/jantimon/favicons-webpack-plugin) - thanks!
- The latest version (4.0.0) of [optimize-css-assets-webpack-plugin](https://github.com/NMFR/optimize-css-assets-webpack-plugin) - thanks!
- You're using both the `FaviconsWebpackPlugin` and `WebpackPwaManifest` to generate icons. This actually generates too many, IMO. I've configured the 2 to play nice together. My relevant options look like:
    WPM icons array:
    ```js
     icons: [
        {
          src: path.resolve('./src/assets/img/favicon-32x32.png'),
          sizes: [40, 48, 58, 60, 72, 80, 87, 120, 152, 180, 167, 512],
          destination: path.join('./', 'assets', 'icons'),
          ios: true
        }, 
        {
          src: path.resolve('./src/assets/img/favicon-32x32.png'),
          size: 1024,
          destination: path.join('./', 'assets', 'icons'),
          ios: 'startup'
        }
      ]
    ```

    FWP icons object:
    ```js
    prefix: 'assets/icons/',
    icons: {
        android: true,
        appleIcon: false,
        appleStartup: false,
        coast: false,
        favicons: true,
        firefox: true,
        opengraph: true,
        twitter: true,
        yandex: false,
        windows: false
      }
    ```
Hope this helps.