@thescientist13 2018-04-02
thanks for the set of eyes on this @DevLab2425 !! 

> I'm using multiple entry files, you have a single. This gives me smaller assets in hopes of avoiding blocking scripts

This is intentional and is considered [best practice now in webpack](https://twitter.com/wSokra/status/969633336732905474) now that CommonsChunk is deprecated and there is the `optimization` configuration option (which I am using in _webpack.config.prod.js_).

In reviewing the linked gist from that tweet, I've actually removed the `cacheGroup` setting since _vendor_ is already a default group.

> I have a minified CSS file injected into index.html. You seem to still have the multiple CSS file issue

Hmm, anyway you could share your config files with me? (maybe a 1:1 DM on slack) 

> You're using both the FaviconsWebpackPlugin and WebpackPwaManifest to generate icons. 

Good call.  Any thoughts on just dropping `FaviconsWebpackPlugin` altogether?